# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.DeviceResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      device_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      device_type:
        example: desktop
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_seen_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      name:
        example: My Desktop
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.FileResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      expires_at:
        example: "2024-02-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mime_type:
        example: application/pdf
        type: string
      name:
        example: document.pdf
        type: string
      size:
        example: 1024000
        type: integer
      storage_path:
        example: users/550e8400-e29b-41d4-a716-446655440000/files/...
        type: string
      storage_type:
        example: local
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.ListDevicesResponse:
    properties:
      devices:
        items:
          $ref: '#/definitions/handlers.DeviceResponse'
        type: array
      total:
        example: 5
        type: integer
    type: object
  handlers.ListFilesResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/handlers.FileResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securePassword123
        type: string
    required:
    - email
    - password
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  handlers.RegisterDeviceRequest:
    properties:
      device_type:
        enum:
        - desktop
        - mobile
        example: desktop
        type: string
      name:
        example: My Desktop
        type: string
    required:
    - device_type
    - name
    type: object
  handlers.RegisterDeviceResponse:
    properties:
      device:
        $ref: '#/definitions/handlers.DeviceResponse'
      device_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securePassword123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  handlers.TurnCredentialsResponse:
    properties:
      password:
        type: string
      realm:
        type: string
      turn_servers:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  handlers.UpdateDeviceRequest:
    properties:
      device_type:
        enum:
        - desktop
        - mobile
        example: desktop
        type: string
      name:
        example: My Updated Desktop
        type: string
    type: object
  handlers.UploadFileResponse:
    properties:
      file_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      storage_path:
        example: users/550e8400-e29b-41d4-a716-446655440000/files/...
        type: string
      uploaded_size:
        example: 1024000
        type: integer
    type: object
  handlers.UserResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@Backend.app
    name: Backend Support
    url: https://Backend.app/support
  description: Backend - SaaS приложение для файлообмена с акцентом на максимальную скорость
    передачи через Local Network First подход
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://Backend.app/terms
  title: Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает JWT токены
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный email или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены используя валидный refresh token.
        Возвращает новую пару токенов.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены успешно обновлены" example:"{\"access_token\":\"eyJhbGci...\",\"refresh_token\":\"eyJhbGci...\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат данных" example:"{\"error\":\"invalid request
            format\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Невалидный или истекший refresh token" example:"{\"error\":\"invalid
            or expired refresh token\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера" example:"{\"error\":\"failed to
            refresh token\"}
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT токены
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь с таким email уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
  /devices:
    get:
      consumes:
      - application/json
      description: Возвращает список всех устройств пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список устройств
          schema:
            $ref: '#/definitions/handlers.ListDevicesResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список устройств
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Регистрирует новое устройство для пользователя и возвращает device_token
        для QR-кода
      parameters:
      - description: Данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Устройство успешно зарегистрировано
          schema:
            $ref: '#/definitions/handlers.RegisterDeviceResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Регистрация устройства
      tags:
      - devices
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет устройство пользователя
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Устройство удалено
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление устройства
      tags:
      - devices
    get:
      consumes:
      - application/json
      description: Возвращает информацию об устройстве по ID
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация об устройстве
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Неверный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение устройства
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Обновляет информацию об устройстве
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Устройство обновлено
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление устройства
      tags:
      - devices
  /devices/{id}/last-seen:
    post:
      consumes:
      - application/json
      description: Обновляет время последней активности устройства (не требует аутентификации)
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Время обновлено
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Неверный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление времени последней активности
      tags:
      - devices
  /files:
    get:
      consumes:
      - application/json
      description: Возвращает список файлов пользователя с пагинацией
      parameters:
      - default: 50
        description: Лимит файлов
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список файлов
          schema:
            $ref: '#/definitions/handlers.ListFilesResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список файлов
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл на сервер через multipart/form-data. Поддерживает
        потоковую загрузку больших файлов.
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Файл успешно загружен
          schema:
            $ref: '#/definitions/handlers.UploadFileResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка файла
      tags:
      - files
  /files/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет файл с сервера
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Файл удален
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление файла
      tags:
      - files
    get:
      consumes:
      - application/json
      description: Возвращает метаданные файла
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Метаданные файла
          schema:
            $ref: '#/definitions/handlers.FileResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Метаданные файла
      tags:
      - files
  /files/{id}/download:
    get:
      consumes:
      - application/json
      description: Скачивает файл с сервера (поддерживает Range requests для частичного
        скачивания). Использует потоковую передачу для больших файлов.
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Смещение в байтах для частичного скачивания
        in: query
        name: offset
        type: integer
      - default: 0
        description: Лимит байт для чтения (0 = весь файл)
        in: query
        name: limit
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл (бинарные данные)
          schema:
            type: file
        "400":
          description: Неверный ID файла" example:"{\"error\":\"file_id is required\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован" example:"{\"error\":\"unauthorized\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу" example:"{\"error\":\"permission denied\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден" example:"{\"error\":\"file not found\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера" example:"{\"error\":\"failed to
            download file\"}
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Скачивание файла
      tags:
      - files
  /webrtc/turn-credentials:
    get:
      consumes:
      - application/json
      description: Возвращает TURN серверы и credentials для WebRTC
      produces:
      - application/json
      responses:
        "200":
          description: TURN credentials
          schema:
            $ref: '#/definitions/handlers.TurnCredentialsResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение TURN credentials
      tags:
      - webrtc
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"